Move test helper to be required by other apps
This is useful to use the metadata as fixture and to be centralised in
one place.
