+++ spec/rails_helper.rb
@@ -6,6 +6,8 @@ require File.expand_path("dummy/config/environment.rb", __dir__)
 abort("The Rails environment is running in production mode!") if Rails.env.production?
 require 'rspec/rails'
 Bundler.require(:development, :test)
+require 'metadata_presenter/test_helpers'
+
 # Add additional requires below this line. Rails is not loaded until this point!
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
@@ -61,5 +63,5 @@ RSpec.configure do |config|
   config.filter_rails_from_backtrace!
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
-  config.include TestHelper
+  config.include MetadataPresenter::TestHelpers
 end
